+++ lib/gem_lookup/flags.rb
@@ -12,7 +12,7 @@ module GemLookup
           help:    { matches: %w[-h --help], desc: 'Display the help screen.' },
           version: { matches: %w[-v --version], desc: 'Display version information.' },
           json:    { matches: %w[-j --json], desc: 'Bulk the output results as raw JSON.' },
-          classic: { matches: %w[-c --classic], desc: 'Stream the output using only words.' }
+          wordy:   { matches: %w[-w --wordy], desc: 'Stream the output using only words.' }
         }
       end