+++ spec/gem_lookup/serializers/wordy_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe GemLookup::Serializers::Classic do
+RSpec.describe GemLookup::Serializers::Wordy do
   describe '.streaming?' do
     it 'returns true' do
       expect(described_class).to be_streaming